requirements.txt

Flask==3.1.1
openai==1.97.0
PyYAML==6.0.2
python-dotenv==1.1.1
